requirements.txt
playwright==1.47.0
pandas==2.2.2
gspread==6.1.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
openpyxl==3.1.5
